Move mock to mock folder in test project
