+++ tests/Fresp.Tests/HttpClientBuilderExtensionsTests.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Fresp.Tests.Mocks;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Fresp.Tests;
